+++ packages/worker/src/hard-clone.ts
