+++ packages/worker/src/index.ts
@@ -1,3 +1,6 @@
 export { hardClone, isShared } from "./hard-clone.js";
-export { rehydrateData, rehydrateVector } from "./rehydrate.js";
+export {
+  rehydrateData,
+  rehydrateVector,
+} from "./rehydrate.js";
 export { preparePostMessage } from "./transferable.js";
